aggregator/requirements.txt
flask
google-cloud-pubsub
google-cloud-logging
gunicorn
kubernetes

